add:添加基于websocket的后台消息推送支持

+++ springMVC-demo/src/main/java/com/celloud/utils/UserUtils.java
@@ -62,7 +62,7 @@ public class UserUtils {
         return true;
     }
 
-    public User deleteById(int id) {
+    public User deleteById(long id) {
         return userMap.remove(id);
     }
